refactor(biobank): split Application abstractions into Export and Repositories
Group the flat Abstractions/ ports into Export/ (IPatientExportSource, ExportParseResult, ExportParseError) and Repositories/ (IBiobankRepository), with matching namespaces.
